рефактор метода generateEnum
