+++ PROTv0.1/Program.cs
@@ -78,7 +78,7 @@ Random n = new Random();
 Generator.GenerateLinear(mas2, 5, 10, false);
 //Generator.GenerateEnum(mas2, 5, 10);
 Question q = new Question("sd", null, 1, "sdas");
-q.
+Question []qq =Generator.GenerateEnum(mas, 5, 10); qq[0].print();
